+++ tests/AccountTest.php
@@ -4,12 +4,14 @@ declare(strict_types=1);
 
 namespace pxgamer\Arionum;
 
+use pxgamer\Arionum\Exceptions\GenericApiException;
+
 final class AccountTest extends TestCase
 {
     /**
      * @test
      * @return void
-     * @throws ApiException
+     * @throws GenericApiException
      */
     public function itGeneratesANewAccount(): void
     {
@@ -22,7 +24,7 @@ final class AccountTest extends TestCase
     /**
      * @test
      * @return void
-     * @throws ApiException
+     * @throws GenericApiException
      */
     public function itGetsTheBalanceForATestAddress(): void
     {
@@ -33,7 +35,7 @@ final class AccountTest extends TestCase
     /**
      * @test
      * @return void
-     * @throws ApiException
+     * @throws GenericApiException
      */
     public function itGetsTheBalanceForATestAlias(): void
     {
@@ -44,7 +46,7 @@ final class AccountTest extends TestCase
     /**
      * @test
      * @return void
-     * @throws ApiException
+     * @throws GenericApiException
      */
     public function itGetsTheBalanceForATestPublicKey(): void
     {
@@ -55,7 +57,7 @@ final class AccountTest extends TestCase
     /**
      * @test
      * @return void
-     * @throws ApiException
+     * @throws GenericApiException
      */
     public function itGetsThePendingBalanceForATestAddress(): void
     {
@@ -66,7 +68,7 @@ final class AccountTest extends TestCase
     /**
      * @test
      * @return void
-     * @throws ApiException
+     * @throws GenericApiException
      */
     public function itGetsTheTransactionsForATestAddress(): void
     {
@@ -78,7 +80,7 @@ final class AccountTest extends TestCase
     /**
      * @test
      * @return void
-     * @throws ApiException
+     * @throws GenericApiException
      */
     public function itGetsTheTransactionsForATestPublicKey(): void
     {
@@ -90,7 +92,7 @@ final class AccountTest extends TestCase
     /**
      * @test
      * @return void
-     * @throws ApiException
+     * @throws GenericApiException
      */
     public function itGetsTheAliasForASpecificAddress(): void
     {
@@ -101,7 +103,7 @@ final class AccountTest extends TestCase
     /**
      * @test
      * @return void
-     * @throws ApiException
+     * @throws GenericApiException
      */
     public function itChecksThatTheAddressIsValid(): void
     {