+++ tests/TransactionTest.php
@@ -4,7 +4,10 @@ declare(strict_types=1);
 
 namespace pxgamer\Arionum;
 
+use pxgamer\Arionum\Models\Transaction;
 use pxgamer\Arionum\Transaction\Version;
+use pxgamer\Arionum\Exceptions\GenericApiException;
+use pxgamer\Arionum\Transaction\TransactionFactory;
 
 final class TransactionTest extends TestCase
 {
@@ -17,7 +20,7 @@ final class TransactionTest extends TestCase
     /**
      * @test
      * @return void
-     * @throws ApiException
+     * @throws GenericApiException
      */
     public function itGetsATransactionByItsId(): void
     {
@@ -28,7 +31,7 @@ final class TransactionTest extends TestCase
     /**
      * @test
      * @return void
-     * @throws ApiException
+     * @throws GenericApiException
      */
     public function itGetsTheNumberOfTransactionsInTheMempool(): void
     {
@@ -42,7 +45,7 @@ final class TransactionTest extends TestCase
      */
     public function itCanGenerateAnAliasSendTransaction(): void
     {
-        $data = Transaction::makeAliasSendInstance(self::TEST_ALIAS, 1.0);
+        $data = TransactionFactory::makeAliasSendInstance(self::TEST_ALIAS, 1.0);
         $this->assertEquals(self::TEST_ALIAS, $data->getDestinationAddress());
         $this->assertEquals(1.0, $data->getValue());
         $this->assertEquals(Version::ALIAS_SEND, $data->getVersion());
@@ -54,7 +57,7 @@ final class TransactionTest extends TestCase
      */
     public function itCanGenerateAnAliasSetTransaction(): void
     {
-        $data = Transaction::makeAliasSetInstance(self::TEST_ADDRESS, self::TEST_ALIAS);
+        $data = TransactionFactory::makeAliasSetInstance(self::TEST_ADDRESS, self::TEST_ALIAS);
         $this->assertEquals(self::TEST_ADDRESS, $data->getDestinationAddress());
         $this->assertEquals(self::TEST_ALIAS, $data->getMessage());
         $this->assertEquals(Version::ALIAS_SET, $data->getVersion());
@@ -67,7 +70,7 @@ final class TransactionTest extends TestCase
      */
     public function itCanGenerateAnMasternodeCreateTransaction(): void
     {
-        $data = Transaction::makeMasternodeCreateInstance(self::TEST_IP, self::TEST_ADDRESS);
+        $data = TransactionFactory::makeMasternodeCreateInstance(self::TEST_IP, self::TEST_ADDRESS);
         $this->assertEquals(self::TEST_ADDRESS, $data->getDestinationAddress());
         $this->assertEquals(self::TEST_IP, $data->getMessage());
         $this->assertEquals(Version::MASTERNODE_CREATE, $data->getVersion());
@@ -81,7 +84,7 @@ final class TransactionTest extends TestCase
      */
     public function itCanGenerateAnMasternodePauseTransaction(): void
     {
-        $data = Transaction::makeMasternodePauseInstance(self::TEST_ADDRESS);
+        $data = TransactionFactory::makeMasternodePauseInstance(self::TEST_ADDRESS);
         $this->assertEquals(self::TEST_ADDRESS, $data->getDestinationAddress());
         $this->assertEquals(Version::MASTERNODE_PAUSE, $data->getVersion());
         $this->assertEquals(Transaction::VALUE_MASTERNODE_COMMAND, $data->getValue());
@@ -94,7 +97,7 @@ final class TransactionTest extends TestCase
      */
     public function itCanGenerateAnMasternodeResumeTransaction(): void
     {
-        $data = Transaction::makeMasternodeResumeInstance(self::TEST_ADDRESS);
+        $data = TransactionFactory::makeMasternodeResumeInstance(self::TEST_ADDRESS);
         $this->assertEquals(self::TEST_ADDRESS, $data->getDestinationAddress());
         $this->assertEquals(Version::MASTERNODE_RESUME, $data->getVersion());
         $this->assertEquals(Transaction::VALUE_MASTERNODE_COMMAND, $data->getValue());
@@ -107,7 +110,7 @@ final class TransactionTest extends TestCase
      */
     public function itCanGenerateAnMasternodeReleaseTransaction(): void
     {
-        $data = Transaction::makeMasternodeReleaseInstance(self::TEST_ADDRESS);
+        $data = TransactionFactory::makeMasternodeReleaseInstance(self::TEST_ADDRESS);
         $this->assertEquals(self::TEST_ADDRESS, $data->getDestinationAddress());
         $this->assertEquals(Version::MASTERNODE_RELEASE, $data->getVersion());
         $this->assertEquals(Transaction::VALUE_MASTERNODE_COMMAND, $data->getValue());